# Optimize Point.NewPointFromBytes